requirements.txt
Flask==3.1.1
google-generativeai==0.8.5
flask-limiter==3.12
flask-talisman==1.1.0
flask-compress==1.17
gunicorn==23.0.0
pytest==8.3.5
pytest-cov==6.1.1
